+++ Cauli/Extensions/NSError+Cauli.swift
@@ -63,3 +63,4 @@ extension NSError {
         }
     }
 }
+// swiftlint:enable nesting