+++ Cauli/Florets/Inspector/Record List/InspectorTableViewDatasource.swift
@@ -26,32 +26,20 @@ internal class InspectorTableViewDatasource: NSObject {
 
     internal private(set) var items: [Record] = [] {
         didSet {
-            filteredItems = self.filteredItems(in: items, with: filter)
+            filteredItems = self.items.filter(self.filter.selects)
         }
     }
     internal private(set) var filteredItems: [Record] = []
-    private var filter: RecordSelector? {
-        didSet {
-            filteredItems = self.filteredItems(in: items, with: filter)
-        }
-    }
     internal var filterString: String? {
         didSet {
-            updateFilter(with: filterString)
+            filteredItems = self.items.filter(self.filter.selects)
         }
     }
-
-    private func filteredItems(in array: [Record], with filter: RecordSelector?) -> [Record] {
-        guard let filter = filter else { return array }
-        return array.filter(filter.selects)
-    }
-
-    private func updateFilter(with filterString: String?) {
+    private var filter: RecordSelector {
         guard let filterString = filterString, !filterString.isEmpty else {
-            filter = nil
-            return
+            return RecordSelector { _ in true }
         }
-        filter = RecordSelector { record in
+        return RecordSelector { record in
             guard let urlString = record.designatedRequest.url?.absoluteString else {
                 return false
             }
@@ -59,8 +47,11 @@ internal class InspectorTableViewDatasource: NSObject {
         }
     }
 
-    private func filter(records: [Record]) -> [Record] {
-        filteredItems(in: records, with: filter)
+    private let formatter: InspectorFloretFormatterType
+
+    init(formatter: InspectorFloretFormatterType) {
+        self.formatter = formatter
+        super.init()
     }
 
     private func performBatchUpdate(in tableView: UITableView, updates: (() -> Void), completion: ((_ finished: Bool) -> Void)? = nil) {
@@ -77,7 +68,7 @@ internal class InspectorTableViewDatasource: NSObject {
     internal func append(records: [Record], to tableView: UITableView, completion: ((_ finished: Bool) -> Void)? = nil) {
         performBatchUpdate(in: tableView, updates: {
             let numberOfExistingRecords = filteredItems.count
-            let numberOfAddedRecords = filter(records: records).count
+            let numberOfAddedRecords = records.filter(filter.selects).count
             let indexPaths = (numberOfExistingRecords..<(numberOfExistingRecords + numberOfAddedRecords)).map { IndexPath(row: $0, section: 0) }
             tableView.insertRows(at: indexPaths, with: .bottom)
             items += records
@@ -107,7 +98,8 @@ extension InspectorTableViewDatasource: UITableViewDataSource {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: InspectorRecordTableViewCell.reuseIdentifier, for: indexPath) as? InspectorRecordTableViewCell else {
             fatalError("Unable to dequeue a cell")
         }
-        cell.configure(with: record(at: indexPath), stringToHighlight: filterString)
+        let data = formatter.listFormattedData(for: record(at: indexPath))
+        cell.configure(with: data, stringToHighlight: filterString)
         cell.accessoryType = .disclosureIndicator
         return cell
     }